+++ src/main.rs
@@ -1,23 +1,16 @@
-use std::process::Command;
-
 use clap::Parser;
 use notify_rust::Notification;
 
 mod package;
-use package::get_package_version;
 
 mod kernel;
-use kernel::KernelInfo;
+use kernel::{KernelChecker, KernelInfo};
 
-/// Parse the output of `xdpyinfo`
-fn parse_xdpyinfo_output(xdpyinfo_output: &str) -> Option<&str> {
-    for line in xdpyinfo_output.lines() {
-        if line.starts_with("X.Org version: ") {
-            return Some(line.split_at(15).1);
-        }
-    }
-    None
-}
+mod checks;
+use checks::{Check, CheckResult};
+mod critical_packages_check;
+use critical_packages_check::CriticalPackagesCheck;
+mod session;
 
 #[derive(Debug, Parser)]
 #[clap(
@@ -26,8 +19,20 @@ fn parse_xdpyinfo_output(xdpyinfo_output: &str) -> Option<&str> {
 )]
 struct Args {
     /// Disable desktop notification
-    #[structopt(long)]
+    #[clap(long)]
     disable_notification: bool,
+
+    /// Comma separated list of packages were we should reboot after an upgrade.
+    #[clap(long, use_delimiter = true, default_value = "systemd,linux-firmware")]
+    reboot_packages: Vec<String>,
+
+    /// Comma separated list of packages were we should restart our session after an upgrade.
+    #[clap(
+        long,
+        use_delimiter = true,
+        default_value = "xorg-server,xorg-xwayland"
+    )]
+    session_restart_packages: Vec<String>,
 }
 
 fn main() {
@@ -38,76 +43,37 @@ fn main() {
         .expect("Could not open pacman database at /var/lib/pacman");
     let db = alpm.localdb();
 
-    // uname output is in the form version-ARCH
-    let output_uname = Command::new("uname")
-        .arg("-r")
-        .output()
-        .expect("Could not execute uname");
-    let output_uname_stdout = String::from_utf8_lossy(&output_uname.stdout);
-    let kernel_info =
-        KernelInfo::from_uname_output(&output_uname_stdout).expect("Failed to parse uname output");
-    let running_kernel_version = kernel_info.version;
+    let kernel_info = KernelInfo::from_uname().expect("Failed to parse uname output");
 
-    let kernel_package = if let Some(variant) = kernel_info.variant {
-        format!("linux-{}", variant)
-    } else {
-        "linux".to_owned()
-    };
+    let mut checkers: Vec<Box<dyn Check>> = vec![Box::new(KernelChecker::new(kernel_info, db))];
 
-    let installed_kernel = get_package_version(&db, &kernel_package)
-        .expect("Could not get version of installed kernel");
+    let session_info = session::SessionInfo::from_utmp();
+    if let Ok(session_info) = session_info {
+        let critical_packages_checker = CriticalPackagesCheck::new(
+            args.reboot_packages,
+            args.session_restart_packages,
+            session_info,
+            db,
+        );
+        checkers.push(Box::new(critical_packages_checker));
+    }
 
-    println!("Kernel");
-    println!(
-        " installed: {} (since {})",
-        installed_kernel.version,
-        installed_kernel.installed_reltime()
-    );
-    println!(" running:   {}", running_kernel_version);
+    let result = checkers
+        .iter()
+        .map(|v| v.check())
+        .max()
+        .expect("No checkers could run");
 
-    let should_reboot = !installed_kernel.version_matches(&running_kernel_version);
-    if should_reboot {
-        println!("You should reboot arch btw!");
+    if result > CheckResult::Nothing {
+        println!("{}", result.summary());
         if !args.disable_notification {
             Notification::new()
-                .summary("Reboot arch btw")
-                .body("Kernel got updated. You should reboot your system!")
+                .summary(result.summary())
+                .body(result.body())
                 .timeout(6000) //milliseconds
                 .show()
                 .map_err(|e| println!("Couldn't send notification: {}", e))
                 .ok();
         }
-    }
-
-    let output_xdpyinfo = Command::new("xdpyinfo")
-        .output()
-        .map_err(|err| println!("Could not execute xdpyinfo: {}", err));
-    if let Ok(output_xdpyinfo) = output_xdpyinfo {
-        let output_xdpyinfo = String::from_utf8_lossy(&output_xdpyinfo.stdout);
-        let running_xorg_version =
-            parse_xdpyinfo_output(&output_xdpyinfo).expect("Could not parse xdpyinfo output");
-        let installed_xorg = get_package_version(&db, "xorg-server")
-            .expect("Could not get version of installed xserver");
-
-        println!("Xorg server");
-        println!(
-            " installed: {} (since {})",
-            installed_xorg.version,
-            installed_xorg.installed_reltime()
-        );
-        println!(" running:   {}", running_xorg_version);
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_parse_xdpyinfo_output() {
-        assert_eq!(
-            Some("1.18.4"),
-            parse_xdpyinfo_output("X.Org version: 1.18.4")
-        );
     }
 }